data/presets/presets: add icons to a few common POI
The end goal would be that when editing an area rich in POI, most of them
should have a different icon so that less point&clicking will be needed
to find the right one to extend.

To achieve this, all common POIs should have individual icons to serve as
mnemonics, and all POIs which are usually located near each other should
usually be assigned different icons.